dags/requirements.txt
apache-airflow==2.7.1
confluent-kafka
requests
